+++ tests/lib/rules-preprocessor/ember_ts/bar.gts
@@ -1,0 +1,5 @@
+export const fortyTwoFromGTS = '42';
+
+<template>
+  {{fortyTwoFromGTS}}
+</template>
