+++ tests/lib/rules-preprocessor/ember_ts/baz.ts
@@ -1,0 +1,1 @@
+export const fortyTwoFromTS = '42';
