+++ tests/lib/rules-preprocessor/tsconfig.eslint.json
@@ -5,6 +5,7 @@
     "strictNullChecks": true
   },
   "include": [
-    "*"
-  ]
+    "**/*.ts",
+    "**/*.gts"
+  ],
 }
